# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Church Records Backend REST API
host: localhost:10120
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /believer:
    x-swagger-router-controller: believer.controller
    get:
      tags: [Believer]
      description: Returns list of believer from Believer Collection
      operationId: getAllBeliever
      responses:
        "200":
          description: Believer Found
          schema:
            type: array
            items:
              $ref: "#/definitions/believer"
        "404":
          description: Empty Collection
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    post:
      tags: [Believer]
      description: Creates new believer entry
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createBeliever
      parameters:
        - name: believerdata
          required: true
          in: body
          description: a new believer details
          schema:
            $ref: "#/definitions/believerreg"
      responses:
        "200":
          description: New Believer Created
          schema:
            $ref: "#/definitions/believer"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Believer]
      description: Remove all believers in database
      operationId: removeAllBeliever
      responses:
        "200":
          description: Successfully removed all believers
          schema:
            $ref: "#/definitions/message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/message"
  /believer/{believerid}:
    x-swagger-router-controller: believer.controller
    get:
      tags: [Believer]
      description: Return single believer by given id
      operationId: getSingleBeliever
      parameters:
        - name: believerid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Believer Found
          schema:
            $ref: "#/definitions/believer"
        "404":
          description: Believer not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Believer]
      description: Remove single believer by given id
      operationId: removeSingleBeliever
      parameters:
        - name: believerid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Successfully removed
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    patch:
      tags: [Believer]
      description: Update by believer id
      operationId: updateSingleBeliever
      parameters:
        - name: believerid
          in: path
          required: true
          type: string
        - name: believer
          in: body
          description: The Updated believer info
          required: true
          schema:
            $ref: "#/definitions/believer"
      responses:
        "200":
          description: Believer Found and Updated
          schema:
            $ref: "#/definitions/believer"
        "404":
          description: Believer not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
#Path to retrieve Believer's info
  /believer/{believerid}/{groupcode}/info:
    x-swagger-router-controller: believer.controller
    get:
      tags: [Believer]
      description: Return infos of a single believer by given by id
      operationId: getBelieverInfo
      parameters:
        - name: believerid
          required: true
          in: path
          type: string
        - name: groupcode
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Info Found
          schema:
            type: array
            items:
              $ref: "#/definitions/info"
        "404":
          description: Believer not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
#path to get believers of certain age range
  /believer/{agefrom}/{ageto}/:
    x-swagger-router-controller: believer.controller
    get:
      tags: [Believer]
      description: Return single believer by given id
      operationId: getRangedBeliever
      parameters:
        - name: agefrom
          required: true
          in: path
          type: number
        - name: ageto
          required: true
          in: path
          type: number
      responses:
        "200":
          description: Believer Found
          schema:
            type: array
            items:
              $ref: "#/definitions/believer"
        "404":
          description: Believer not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# Add Beliver's relative Path
  /believer/{believerid}/addrelative:
    x-swagger-router-controller: believer.controller
    post:
      tags: [Believer]
      description: Add relative to the believer
      consumes:
        - application/json
      produces:
        - application/json
      operationId: addRelative
      parameters:
        - name: believerid
          required: true
          in: path
          type: string
        - name: relativedata
          required: true
          in: body
          description: new relative data
          schema:
            $ref: "#/definitions/relative"
      responses:
        "200":
          description: New relative Added
          schema:
            $ref: "#/definitions/message"
        "404":
          description: Invalid Believer
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# Admin Path Definition
...
  /believer/{believerid}/updateprofilepic:
    x-swagger-router-controller: believer.controller
    post:
      tags: [Believer]
      description: Updates Profilepic
      operationId: updateProfilePic
      parameters:
        - name: believerid
          in: path
          required: true
          type: string
        - name: profiledata
          in: body
          description: The Updated info
          required: true
          schema:
            $ref: "#/definitions/profiledata"
      responses:
        "200":
          description: Believer Found and Updated
          schema:
            $ref: "#/definitions/message"
        "404":
          description: Believer not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# Path Defition to update Notification Token
  /believer/{believerid}/updatetoken:
    x-swagger-router-controller: believer.controller
    post:
      tags: [Believer]
      description: Updates token
      operationId: updateDeviceId
      parameters:
        - name: believerid
          in: path
          required: true
          type: string
        - name: device_id
          in: body
          description: The Updated info
          required: true
          schema:
            $ref: "#/definitions/tokendata"
      responses:
        "200":
          description: Believer Found and Updated
          schema:
            $ref: "#/definitions/message"
        "404":
          description: Believer not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The Start of Dio Marchavel
...
definitions:
...
  believer:
    properties:
      _id:
        type: string
      believercode:
        type: string
      title:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communityid:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      kipaimarastatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
#Relative model Definition
  relative:
    properties:
      relation:
        type: string
      person:
        type: string
# Mode Definition for profilepic
  profiledata:
    required:
      - believerinfo
    properties:
      believerinfo:
        type: string
# Mode Definition for token
  tokendata:
    required:
      - device_id
    properties:
      device_id:
        type: string
# Believer Model Definition
  believerreg:
    required:
      - believercode
      - firstname
      - lastname
      - status
      - password
      - title
      - gender
      - dob
      - residenceadress
      - regdate
      - baptismstatus
      - churchcode
      - churchname
      - diocode
      - dioname
    properties:
      _id:
        type: string
      believercode:
        type: string
      password:
        type: string
      title:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communityid:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      kipaimarastatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
# Admin reg Model Definition
...
  message:
    required:
      - success
      - message
    properties:
      success:
        type: boolean
      message:
        type: string
# Info Model Definition
  info:
    required:
      - infotitle
      - groupcode
      - body
      - ageto
      - agefrom
    properties:
      _id:
        type: string
      infotitle:
        type: string
      groupcode:
        type: string
      agefrom:
        type: number
      ageto:
        type: number
      body:
        type: string
      valid:
        type: number
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
# Country model Definition
...